Add sink implementation
